+++ common/cosmwasm-smart-contracts/service-provider-directory/src/lib.rs
@@ -5,3 +5,5 @@ pub mod types;
 
 // Re-export all types at the top-level
 pub use types::*;
+
+pub use cosmwasm_std::{Addr, Coin, Decimal, Fraction};