+++ nym-wallet/nym-wallet-types/src/network.rs
@@ -114,6 +114,8 @@ mod sandbox {
         "nymt1k8re7jwz6rnnwrktnejdwkwnncte7ek7kk6fvg";
     pub(crate) const COCONUT_DKG_CONTRACT_ADDRESS: &str =
         "nymt1k8re7jwz6rnnwrktnejdwkwnncte7ek7kk6fvg";
+    pub(crate) const SERVICE_PROVIDER_DIRECTORY_CONTRACT_ADDRESS: &str =
+        "n1ryt076cufyddallg5x0gz3qjz0pd3wg0m4cwkg9njhmlnp6u88qq6nczgj";
     pub(crate) const _ETH_CONTRACT_ADDRESS: [u8; 20] =
         hex_literal::hex!("8e0DcFF7F3085235C32E845f3667aEB3f1e83133");
     pub(crate) const _ETH_ERC20_CONTRACT_ADDRESS: [u8; 20] =
@@ -149,6 +151,9 @@ mod sandbox {
                 group_contract_address: parse_optional_str(GROUP_CONTRACT_ADDRESS),
                 multisig_contract_address: parse_optional_str(MULTISIG_CONTRACT_ADDRESS),
                 coconut_dkg_contract_address: parse_optional_str(COCONUT_DKG_CONTRACT_ADDRESS),
+                service_provider_directory_contract_address: parse_optional_str(
+                    SERVICE_PROVIDER_DIRECTORY_CONTRACT_ADDRESS,
+                ),
             },
         }
     }
@@ -176,6 +181,8 @@ mod qa {
     pub(crate) const MULTISIG_CONTRACT_ADDRESS: &str = "n17p9rzwnnfxcjp32un9ug7yhhzgtkhvl988qccs";
     pub(crate) const COCONUT_DKG_CONTRACT_ADDRESS: &str =
         "n17p9rzwnnfxcjp32un9ug7yhhzgtkhvl988qccs";
+    pub(crate) const SERVICE_PROVIDER_DIRECTORY_CONTRACT_ADDRESS: &str =
+        "n1ryt076cufyddallg5x0gz3qjz0pd3wg0m4cwkg9njhmlnp6u88qq6nczgj";
     pub(crate) const _ETH_CONTRACT_ADDRESS: [u8; 20] =
         hex_literal::hex!("0000000000000000000000000000000000000000");
     pub(crate) const _ETH_ERC20_CONTRACT_ADDRESS: [u8; 20] =
@@ -210,6 +217,9 @@ mod qa {
                 group_contract_address: parse_optional_str(GROUP_CONTRACT_ADDRESS),
                 multisig_contract_address: parse_optional_str(MULTISIG_CONTRACT_ADDRESS),
                 coconut_dkg_contract_address: parse_optional_str(COCONUT_DKG_CONTRACT_ADDRESS),
+                service_provider_directory_contract_address: parse_optional_str(
+                    SERVICE_PROVIDER_DIRECTORY_CONTRACT_ADDRESS,
+                ),
             },
         }
     }