+++ nym-wallet/src/context/main.tsx
@@ -17,6 +17,7 @@ import {
   switchAccount,
 } from '../requests';
 import { Console } from '../utils/console';
+import { createSignInWindow, getReactState, setReactState } from '../requests/app';
 import { toDisplay } from '../utils';
 
 export const urls = (networkName?: Network) =>
@@ -63,9 +64,18 @@ export type TAppContext = {
   handleShowTerminal: () => void;
   signInWithPassword: (password: string) => void;
   logOut: () => void;
+<<<<<<< HEAD
+=======
+  keepState: () => Promise<void>;
+>>>>>>> master
   printBalance: string;
   printVestedBalance?: string; // spendable vested token
 };
+
+interface RustState {
+  network?: Network;
+  loginType?: 'mnemonic' | 'password';
+}
 
 export const AppContext = createContext({} as TAppContext);
 
@@ -91,6 +101,28 @@ export const AppProvider: FCWithChildren = ({ children }) => {
   const userBalance = useGetBalance(clientDetails);
   const navigate = useNavigate();
   const { enqueueSnackbar } = useSnackbar();
+
+  const initFromRustState = async () => {
+    const stateJson = await getReactState();
+    if (stateJson) {
+      const state: RustState = JSON.parse(stateJson);
+      setNetwork(state.network);
+      setLoginType(state.loginType);
+    }
+  };
+
+  useEffect(() => {
+    initFromRustState();
+  }, []);
+
+  const keepState = async () => {
+    // add any state from this context to store in the Rust process
+    const state: RustState = {
+      network,
+      loginType,
+    };
+    setReactState(JSON.stringify(state));
+  };
 
   const clearState = () => {
     userBalance.clearAll();
@@ -227,9 +259,16 @@ export const AppProvider: FCWithChildren = ({ children }) => {
   };
 
   const logOut = async () => {
-    await signOut();
-    setClientDetails(undefined);
-    enqueueSnackbar('Successfully logged out', { variant: 'success' });
+    setIsLoading(true);
+    try {
+      await signOut();
+      await setReactState(undefined);
+      setClientDetails(undefined);
+      enqueueSnackbar('Successfully logged out', { variant: 'success' });
+      await createSignInWindow();
+    } finally {
+      setIsLoading(false);
+    }
   };
 
   const onAccountChange = async ({ accountId, password }: { accountId: string; password: string }) => {
@@ -284,6 +323,7 @@ export const AppProvider: FCWithChildren = ({ children }) => {
       handleShowTerminal,
       logIn,
       logOut,
+      keepState,
       onAccountChange,
       showSendModal,
       showReceiveModal,